requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
aiofiles
python-jose[cryptography]
bcrypt
itsdangerous
